requirements.txt
altgraph==0.17
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.20.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dlib==19.22.0
dnspython==1.16.0
docopt==0.6.2
eventlet==0.31.0
face-recognition==1.3.0
face-recognition-models==0.3.0
future==0.18.2
greenlet==1.1.0
gTTS==2.2.2
hog==0.1.7
idna==2.10
imageio==2.9.0
imbalanced-learn==0.8.0
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
Js2Py==0.71
kiwisolver==1.3.1
ldap3==2.9
matplotlib==3.4.1
matplotlib-inline==0.1.2
networkx==2.5.1
numpy==1.21.0
opencv-contrib-python==3.4.13.47
opencv-python==4.5.1.48
packaging==20.9
parso==0.8.2
pefile==2021.5.24
pickleshare==0.7.5
Pillow==8.2.0
pipwin==0.5.1
playsound==1.2.2
prompt-toolkit==3.0.19
pyasn1==0.4.8
PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pydirectory==0.1.15
Pygments==2.9.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.3
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.0.3
pySmartDL==1.3.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
threadpoolctl==2.1.0
tifffile==2021.4.8
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.4
wcwidth==0.2.5
wget==3.2
